feat: add ability to pack many plugins, change artefacts paths

+++ src/helpers/build/mod.rs
@@ -26,9 +26,12 @@ pub struct Params {}
 
 pub fn main(_params: &Params) {
     let out_dir = get_output_path();
-    let out_manifest_path = Path::new(&out_dir).join("manifest.yaml");
     let pkg_version = env::var("CARGO_PKG_VERSION").unwrap();
     let pkg_name = env::var("CARGO_PKG_NAME").unwrap();
+    let out_manifest_path = Path::new(&out_dir)
+        .join(&pkg_name)
+        .join(&pkg_version)
+        .join("manifest.yaml");
     let plugin_path = out_dir.join(&pkg_name).join(&pkg_version);
     let lib_name = format!("lib{}.{LIB_EXT}", pkg_name.replace('-', "_"));
 
@@ -64,45 +67,58 @@ pub fn main(_params: &Params) {
     let crate_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
     let crate_dir = Path::new(&crate_dir);
 
-    let template_path = crate_dir.join(MANIFEST_TEMPLATE_NAME);
-    let template =
-        fs::read_to_string(template_path).expect("template for manifest plugin not found");
-    let template = liquid::ParserBuilder::with_stdlib()
-        .build()
-        .unwrap()
-        .parse(&template)
-        .expect("invalid manifest template");
-
     let migrations_dir = crate_dir.join("migrations");
-    let migrations: Vec<String> = match fs::read_dir(&migrations_dir) {
-        Ok(dir) => dir
-            .map(|path| {
-                path.unwrap()
-                    .path()
-                    .strip_prefix(crate_dir)
-                    .unwrap()
-                    .to_string_lossy()
-                    .into()
-            })
-            .collect(),
-        Err(_) => Vec::new(),
-    };
-
-    let template_ctx = liquid::object!({
-        "version": pkg_version,
-        "migrations": migrations,
-    });
-
-    fs::write(&out_manifest_path, template.render(&template_ctx).unwrap()).unwrap();
+    let mut migrations: Vec<String> = vec![];
+    if migrations_dir.exists() {
+        migrations = match fs::read_dir(&migrations_dir) {
+            Ok(dir) => dir
+                .map(|path| {
+                    path.unwrap()
+                        .path()
+                        .strip_prefix(crate_dir)
+                        .unwrap()
+                        .to_string_lossy()
+                        .into()
+                })
+                .collect(),
+            Err(_) => Vec::new(),
+        };
+    }
 
     // Copy migrations directory and manifest into newest plugin version
     if !migrations.is_empty() {
         let mut cp_opts = CopyOptions::new();
         cp_opts.overwrite = true;
-        dir::copy(&migrations_dir, &out_dir, &cp_opts).unwrap();
-        dir::copy(out_dir.join("migrations"), &plugin_path, &cp_opts).unwrap();
+        dir::copy(
+            &migrations_dir,
+            out_dir.join(&pkg_name).join(&pkg_version),
+            &cp_opts,
+        )
+        .unwrap();
     }
-    fs::copy(out_manifest_path, plugin_path.join("manifest.yaml")).unwrap();
+
+    if crate_dir.join(MANIFEST_TEMPLATE_NAME).exists() {
+        let template_path = crate_dir.join(MANIFEST_TEMPLATE_NAME);
+        let template =
+            fs::read_to_string(template_path).expect("template for manifest plugin not found");
+        let template = liquid::ParserBuilder::with_stdlib()
+            .build()
+            .unwrap()
+            .parse(&template)
+            .expect("invalid manifest template");
+
+        let template_ctx = liquid::object!({
+            "version": pkg_version,
+            "migrations": migrations,
+        });
+
+        fs::write(&out_manifest_path, template.render(&template_ctx).unwrap()).unwrap();
+    } else {
+        log::warn!(
+            "Couldn't find manifest.yaml template at {}, skipping its generation...",
+            crate_dir.display()
+        );
+    }
 
     // Create symlinks for newest plugin version, which would be created after build.rs script
     std::os::unix::fs::symlink(out_dir.join(&lib_name), plugin_path.join(lib_name)).unwrap();
